now it compresses any file
- but when it unzips, the output is wrong

+++ Huffman/ByteFrequency.cpp
@@ -10,7 +10,7 @@ ByteFrequency::ByteFrequency(int f, unsigned char b) : frequency(f), byte_code(b
 
 ByteFrequency::ByteFrequency() : frequency(0), byte_code(0) {}
 
-short int ByteFrequency::get_frequency() const {
+int ByteFrequency::get_frequency() const {
     return frequency;
 }
 
@@ -41,7 +41,7 @@ bool ByteFrequency::is_empty()
 ofstream& operator<<(ofstream& os, const ByteFrequency& bf)
 {
     os << bf.get_byte_code();
-    short int a = bf.get_frequency();
+    short int a = (short int)bf.get_frequency();
     os.write(reinterpret_cast<const char*>(&a), sizeof(a));
     return os;
 }
